Rename domain-cache-worker to protein-cache-worker
